+++ src/peer_manager.rs
@@ -303,6 +303,7 @@ impl PeerManager {
                 dynamic_neighbors: Vec::new(),
                 evpn_instances: Vec::new(),
                 ethernet_segments: Vec::new(),
+                apply_bum_enforcement: false,
             },
         )
     }
@@ -2167,6 +2168,7 @@ mod tests {
             file_path: None,
             evpn_instances: Vec::new(),
             ethernet_segments: Vec::new(),
+            apply_bum_enforcement: false,
         }
     }